+++ catapult/__init__.py
@@ -27,6 +27,7 @@ gi.require_version("Gio", "2.0")
 gi.require_version("GObject", "2.0")
 gi.require_version("Gtk", "4.0")
 gi.require_version("Pango", "1.0")
+gi.require_version("PangoCairo", "1.0")
 
 from gi.repository import GLib
 from pathlib import Path