+++ catapult/test/test_util.py
@@ -47,6 +47,9 @@ class TestUtil(catapult.test.TestCase):
     def test_get_monitor(self):
         assert catapult.util.get_monitor()
 
+    def test_get_scale_factor(self):
+        assert catapult.util.get_scale_factor() >= 1
+
     def test_get_screen_size(self):
         width, height = catapult.util.get_screen_size()
         assert width and height