+++ NEWS.md
@@ -1,3 +1,11 @@
+PENDING: Catapult 1.2
+=====================
+
+* Add characters plugin to copy characters and emojis
+* Add dependency on Cairo
+* Add dependency on Noto Sans font
+* Add dependency on Noto Color Emoji font
+
 2025-01-07: Catapult 1.1.1
 ==========================